only sends geolocation response when user has allowed access
